Software Design Specification
The purpose of this document is to provide a high-level design framework around which we’d build our Code Utility Web Application. The document also provides a list of requirements against which we’d be testing our final project and determine whether we were able to successfully implement the application according to design.