Software Requirement Specification (SRS)
We all go through the trouble of googling about any function, data type or syntax while coding while learning a new programming language. Programming Assistant is a Web App where the user can search a keyword/task and get all the help about it with a code snippet and description, in whichever language user chooses from the ones provided by the system. The user can even ask questions like how can we validate a password and get the description and code snippet for the same. It can be used by him/her while coding! The search time is also reduced as when you search on Google, it provides references to multiple sites to get a single working solution!
This app also provides the user (who must be logged in) with the option to save all their codes and functions made by them which will work as a guide for them in future. The user can even share his snippets with his friends as well as publically (i.e.
making it visible to all registered users).